removed unused page, made landing page of the sbc app the submit page
